Update Canny testbenches for current dataflow behaviour

+++ testbench/nms_reference.h
@@ -2,13 +2,11 @@
 #define NMS_REFERENCE_H
 
 #include <cstdint>
-
-#include "config.h"
-#include "canny_types.h"
+#include "canny_stages.h"
 
 void non_maximum_suppression_reference(
     const GradientPixel input[HEIGHT * WIDTH],
-    std::uint8_t output[HEIGHT * WIDTH]
+    std::uint16_t output[HEIGHT * WIDTH]
 );
 
 #endif
